Fixes #2: Particles that are only hidden in the viewport don't render when the add-on is on
Seems that the frame_change_post handler is triggered when rendering, which eventually causes the current frame to be set to something new, which conflicts with the render.

Now, rendering is detected with render_pre and render_post handlers so the frame_change_post handler can be skipped when rendering.